Sub management almost finished, tidy up some methods

+++ src/main/java/bdproject/controller/gui/adminarea/AdminSubManagementController.java
@@ -5,6 +5,7 @@ import bdproject.controller.gui.ViewController;
 import bdproject.model.Queries;
 import bdproject.tables.pojos.*;
 import bdproject.utils.FXUtils;
+import bdproject.utils.LocaleUtils;
 import bdproject.view.StringRepresentations;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -15,6 +16,7 @@ import javafx.scene.control.*;
 import javafx.stage.Stage;
 
 import javax.sql.DataSource;
+import java.math.BigDecimal;
 import java.net.URL;
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -33,23 +35,29 @@ public class AdminSubManagementController extends AbstractViewController impleme
     private static final int NO_CLIENT = -1;
 
     private List<Contratti> subsList = Collections.emptyList();
+    private List<TipologieUso> useTypes;
     private final Map<String, BiPredicate<Contratti, Connection>> statuses = Map.of(
             "Attivo", (s, c) -> true,
             "Interrotto", (s, c) -> s.getDatacessazione() == null && Queries.hasOngoingInterruption(s, c),
             "Cessato", (s, c) -> s.getDatacessazione() != null
     );
 
-    private final DateTimeFormatter df_it = DateTimeFormatter.ofPattern("dd/MM/yyyy").withLocale(Locale.ITALIAN);
+    private final DateTimeFormatter dateIt = LocaleUtils.getItDateFormatter();
 
+    @FXML private Button back;
     @FXML private TableView<Contratti> subsTable;
     @FXML private TableColumn<Contratti, Integer> clientIdCol;
     @FXML private TableColumn<Contratti, Integer> subIdCol;
     @FXML private TableColumn<Contratti, String> zoneCol;
     @FXML private TableColumn<Contratti, Integer> planIdCol;
+    @FXML private TableColumn<Contratti, String> incomeDiscountCol;
     @FXML private TextField clientIdFilter;
     @FXML private ComboBox<String> statusFilter;
     @FXML private CheckBox lastReportFilter;
     @FXML private CheckBox pastDeadlineFilter;
+    @FXML private Button interruptButton;
+    @FXML private TextField interruptionDescription;
+    @FXML private Button reactivateButton;
 
     @FXML private TableView<Contratti> requestTable;
     @FXML private TableColumn<Contratti, Integer> reqClientIdCol;
@@ -58,9 +66,19 @@ public class AdminSubManagementController extends AbstractViewController impleme
     @FXML private TableColumn<Contratti, String> reqActivationCol;
 
     @FXML private TableView<Letture> measurementsTable;
+    @FXML private TableColumn<Letture, String> measPublishDateCol;
+    @FXML private TableColumn<Letture, String> measConsumptionCol;
+    @FXML private TableColumn<Letture, String> measConfirmedCol;
+    @FXML private Button confirmMeasurement;
+
     @FXML private TableView<Bollette> reportsTable;
-    @FXML private Button back;
-    @FXML private Button interruptButton;
+    @FXML private TableColumn<Bollette, String> repPublishDateCol;
+    @FXML private TableColumn<Bollette, String> repDeadlineCol;
+    @FXML private TableColumn<Bollette, String> repPaidDateCol;
+    @FXML private TableColumn<Bollette, String> repEstimatedCol;
+    @FXML private TableColumn<Bollette, String> repConsumptionCol;
+    @FXML private TableColumn<Bollette, String> repActivationCostCol;
+
     @FXML private Button publishReport;
 
 
@@ -74,12 +92,18 @@ public class AdminSubManagementController extends AbstractViewController impleme
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
+        try (Connection conn = getDataSource().getConnection()) {
+            useTypes = Queries.getUsages(conn);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+
         populateStatusFilter();
         initSubsTable();
-        doRefreshSubs();
-
         initRequestTable();
-        doRefreshRequests();
+        initMeasurementsTable();
+        initReportsTable();
+        refreshAll();
     }
 
     private void populateStatusFilter() {
@@ -92,11 +116,22 @@ public class AdminSubManagementController extends AbstractViewController impleme
     }
 
     private void initSubsTable() {
-        subsTable.setItems(FXCollections.observableList(subsList));
-
+        subsTable.getSelectionModel().selectedItemProperty().addListener((obs, oldS, newS) -> {
+            if (newS != null && oldS != newS) {
+                doRefreshMeasurements();
+                doRefreshReports();
+            }
+        });
         clientIdCol.setCellValueFactory(c -> new SimpleIntegerProperty(c.getValue().getCodicecliente()).asObject());
         subIdCol.setCellValueFactory(c -> new SimpleIntegerProperty(c.getValue().getIdcontratto()).asObject());
         planIdCol.setCellValueFactory(c -> new SimpleIntegerProperty(c.getValue().getCodiceofferta()).asObject());
+        incomeDiscountCol.setCellValueFactory(c -> {
+            final boolean hasDiscount = useTypes
+                    .stream()
+                    .filter(u -> u.getNome().equals(c.getValue().getTipouso()))
+                    .anyMatch(u -> u.getScontoreddito() == 1);
+            return new SimpleStringProperty(hasDiscount ? "Sì" : "No");
+        });
 
         zoneCol.setCellValueFactory(c -> {
             final Zone zone = Queries.getZone(c.getValue(), getDataSource()).orElseThrow();
@@ -133,10 +168,68 @@ public class AdminSubManagementController extends AbstractViewController impleme
         subsTable.setItems(FXCollections.observableList(subsList));
     }
 
+    @FXML
+    private void doInterrupt() {
+        final Contratti sub = subsTable.getSelectionModel().getSelectedItem();
+        if (sub != null) {
+            if (!interruptionDescription.getText().equals("")) {
+                try (Connection conn = getDataSource().getConnection()) {
+                    final int result = Queries.interruptSubscription(
+                            sub.getIdcontratto(), interruptionDescription.getText(), conn);
+                    if (result != 0) {
+                        FXUtils.showBlockingWarning("Contratto interrotto.");
+                        interruptionDescription.setText("");
+                        refreshAll();
+                    } else {
+                        FXUtils.showBlockingWarning("Impossibile interrompere il contratto.");
+                    }
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                    FXUtils.showError(e.getMessage());
+                }
+            } else {
+                FXUtils.showBlockingWarning("Inserisci una motivazione.");
+            }
+        } else {
+            FXUtils.showBlockingWarning("Seleziona un contratto.");
+        }
+    }
+
+    @FXML
+    private void doReactivate() {
+        final Contratti sub = subsTable.getSelectionModel().getSelectedItem();
+        if (sub != null) {
+            try (Connection conn = getDataSource().getConnection()) {
+                final int result = Queries.reactivateSubscription(sub.getIdcontratto(), conn);
+                if (result != 0) {
+                    FXUtils.showBlockingWarning("Contratto riattivato.");
+                    refreshAll();
+                } else {
+                    FXUtils.showBlockingWarning("Impossibile riattivare il contratto.");
+                }
+            } catch (SQLException e) {
+                e.printStackTrace();
+                FXUtils.showError(e.getMessage());
+            }
+        } else {
+            FXUtils.showBlockingWarning("Seleziona un contratto.");
+        }
+    }
+
+    @FXML
+    private void showSubDetails() {
+        final Contratti selected = subsTable.getSelectionModel().getSelectedItem();
+        if (selected != null) {
+            createSubWindow(AdminSubDetailsController.create(null, getDataSource(), selected));
+        } else {
+            FXUtils.showBlockingWarning("Seleziona un contratto.");
+        }
+    }
+
     private void initRequestTable() {
         reqClientIdCol.setCellValueFactory(c -> new SimpleIntegerProperty(c.getValue().getCodicecliente()).asObject());
         reqSubIdCol.setCellValueFactory(c -> new SimpleIntegerProperty(c.getValue().getIdcontratto()).asObject());
-        reqRequestDateCol.setCellValueFactory(c -> new SimpleStringProperty(df_it.format(c.getValue().getDatarichiesta())));
+        reqRequestDateCol.setCellValueFactory(c -> new SimpleStringProperty(dateIt.format(c.getValue().getDatarichiesta())));
         reqActivationCol.setCellValueFactory(c -> new SimpleStringProperty(c.getValue().getNomeattivazione()));
     }
 
@@ -165,7 +258,7 @@ public class AdminSubManagementController extends AbstractViewController impleme
     private void doActivateSub() {
         final var selectedRequest = requestTable.getSelectionModel().getSelectedItem();
         if (selectedRequest != null) {
-            FXUtils.showYesNoDialog("Vuoi davvero attivare questa fornitura?", () -> {
+            FXUtils.showConfirmationDialog("Vuoi attivare questa fornitura?", () -> {
                 try (Connection conn = getDataSource().getConnection()) {
                     Queries.activateSub(selectedRequest.getIdcontratto(), conn);
                     FXUtils.showBlockingWarning("Contratto attivato.");
@@ -182,7 +275,7 @@ public class AdminSubManagementController extends AbstractViewController impleme
     private void doDeleteRequest() {
         final var selectedRequest = requestTable.getSelectionModel().getSelectedItem();
         if (selectedRequest != null) {
-            FXUtils.showYesNoDialog("Vuoi davvero eliminare definitivamente questa richiesta?", () -> {
+            FXUtils.showConfirmationDialog("Vuoi davvero eliminare definitivamente questa richiesta?", () -> {
                 try (Connection conn = getDataSource().getConnection()) {
                     Queries.deleteRequest(selectedRequest.getIdcontratto(), conn);
                     FXUtils.showBlockingWarning("Richiesta eliminata.");
@@ -195,23 +288,80 @@ public class AdminSubManagementController extends AbstractViewController impleme
         }
     }
 
+    private void initMeasurementsTable() {
+        measPublishDateCol.setCellValueFactory(c -> new SimpleStringProperty(dateIt.format(c.getValue().getDataeffettuazione())));
+        measConsumptionCol.setCellValueFactory(c -> new SimpleStringProperty(c.getValue().getConsumi().toString()));
+        measConfirmedCol.setCellValueFactory(c -> new SimpleStringProperty(c.getValue().getConfermata() == 1 ? "Sì" : "No"));
+    }
+
+    private void doRefreshMeasurements() {
+        final Contratti sub = subsTable.getSelectionModel().getSelectedItem();
+        if (sub != null) {
+            try (Connection conn = getDataSource().getConnection()) {
+                final List<Letture> measurements = Queries.fetchMeasurements(sub.getIdcontratto(), conn);
+                measurementsTable.setItems(FXCollections.observableList(measurements));
+            } catch (SQLException e) {
+                e.printStackTrace();
+                FXUtils.showError("Impossibile aggiornare la tabella delle letture.");
+            }
+        } else {
+            measurementsTable.setItems(FXCollections.emptyObservableList());
+        }
+    }
+
+    @FXML
+    private void doConfirmMeasurement() {
+        final Letture selectedMeas = measurementsTable.getSelectionModel().getSelectedItem();
+        if (selectedMeas != null) {
+            try (Connection conn = getDataSource().getConnection()) {
+                final int result = Queries.confirmMeasurement(selectedMeas.getContatore(), selectedMeas.getDataeffettuazione(), conn);
+                if (result != 0) {
+                    FXUtils.showBlockingWarning("Lettura confermata.");
+                } else {
+                    FXUtils.showBlockingWarning("Lettura già confermata.");
+                }
+            } catch (SQLException e) {
+                e.printStackTrace();
+                FXUtils.showBlockingWarning(StringRepresentations.getGenericError());
+            }
+        } else {
+            FXUtils.showBlockingWarning("Seleziona una lettura.");
+        }
+    }
+
+    private void initReportsTable() {
+        repPublishDateCol.setCellValueFactory(c -> new SimpleStringProperty(dateIt.format(c.getValue().getDataemissione())));
+        repDeadlineCol.setCellValueFactory(c -> new SimpleStringProperty(dateIt.format(c.getValue().getDatascadenza())));
+        repPaidDateCol.setCellValueFactory(c -> new SimpleStringProperty(dateIt.format(c.getValue().getDatapagamento())));
+        repEstimatedCol.setCellValueFactory(c -> new SimpleStringProperty(c.getValue().getStimata() == 1 ? "Sì" : "No"));
+        repConsumptionCol.setCellValueFactory(c -> new SimpleStringProperty(c.getValue().getConsumi().toString()));
+        repActivationCostCol.setCellValueFactory(c -> new SimpleStringProperty(c.getValue().getCostoattivazione().toString()));
+    }
+
+    private void doRefreshReports() {
+        final Contratti sub = subsTable.getSelectionModel().getSelectedItem();
+        if (sub != null) {
+            try (Connection conn = getDataSource().getConnection()) {
+                final var reports = Queries.getReports(sub, conn);
+                reportsTable.setItems(FXCollections.observableList(reports));
+            } catch (SQLException e) {
+                e.printStackTrace();
+                FXUtils.showBlockingWarning("Impossibile aggiornare la tabella delle bollette.");
+            }
+        } else {
+            reportsTable.setItems(FXCollections.emptyObservableList());
+        }
+    }
+
     private void refreshAll() {
         doRefreshSubs();
         doRefreshRequests();
+        doRefreshMeasurements();
+        doRefreshReports();
     }
 
     @FXML
     private void goBack() {
         switchTo(AdminChooseAreaController.create(getStage(), getDataSource()));
-    }
-
-    @FXML
-    private void showSubDetails() {
-        final Contratti selected = subsTable.getSelectionModel().getSelectedItem();
-        if (selected != null) {
-            createSubWindow(AdminSubDetailsController.create(null, getDataSource(), selected));
-        } else {
-            FXUtils.showBlockingWarning("Seleziona un contratto.");
-        }
     }
 }
